+++ sndata/sdss/__init__.py
@@ -7,7 +7,7 @@ spectroscopic components of the Sako et al. 2018 data release is provided
 by separate modules. For more information on SDSS see https://www.sdss.org .
 """
 
-from . import sako18, sako18spec
+from ._sako18 import Sako18
 
 survey_name = 'Sloan Digital Sky Survey'
 survey_abbrev = 'SDSS'
